WIP: Move _resolve_crawl_refs and states to basecrawls
